Changes per review.
